% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{iQualityR.doe-package}
\alias{iQualityR.doe}
\alias{iQualityR.doe-package}
\title{iQualityR.doe}
\description{
Design of Experiments tools for the iQualityR ecosystem.
}
\author{
\strong{Maintainer}: iQualityR Team \email{iQualityR@example.com}

}
\keyword{internal}
